More pre-review fixes
